Create named.conf
We used Bind9 DNS server and set it as the primary server of our LAN. We configured it to respond to the DNS queries from the hubs with our machine’s address instead of the heartbeat server. This is the main configuration file of Bind9. It defines the correct zone files for each domain.